+++ marshy/factory/marshaller_factory_abc.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 from functools import total_ordering
 from typing import Optional, Type
 
-_MarshallerContext = "marshy.marshaller_context.MarshallerContext"
+_MarshyContext = "marshy.marshy_context.MarshyContext"
 _MarshallerABC = "marshy.marshaller_abc.MarshallerABC"
 
 
@@ -12,7 +12,7 @@ class MarshallerFactoryABC(ABC):
 
     @abstractmethod
     def create(
-        self, context: _MarshallerContext, type_: Type
+        self, context: _MarshyContext, type_: Type
     ) -> Optional[_MarshallerABC]:
         """
         Create a new marshaller instance if possible - if not return None
