+++ marshy/factory/optional_marshaller_factory.py
@@ -4,10 +4,10 @@ from typing import Type, Optional, Union
 
 import typing_inspect
 
-from marshy import marshaller_context
 from marshy.factory.marshaller_factory_abc import MarshallerFactoryABC
 from marshy.marshaller import marshaller_abc
 from marshy.marshaller.optional_marshaller import OptionalMarshaller
+from marshy.marshy_context import MarshyContext
 
 
 @dataclass
@@ -15,7 +15,7 @@ class OptionalMarshallerFactory(MarshallerFactoryABC):
     priority: int = 100
 
     def create(
-        self, context: marshaller_context.MarshallerContext, type_: Type
+        self, context: MarshyContext, type_: Type
     ) -> Optional[marshaller_abc.MarshallerABC]:
         origin = typing_inspect.get_origin(type_)
         if origin == Union:
