+++ marshy/marshaller/deferred_marshaller.py
@@ -3,7 +3,7 @@ from typing import TypeVar
 
 from marshy import ExternalType
 from marshy.marshaller.marshaller_abc import MarshallerABC
-from marshy.marshaller_context import MarshallerContext
+from marshy.marshy_context import MarshyContext
 
 T = TypeVar("T")
 
@@ -11,7 +11,7 @@ T = TypeVar("T")
 @dataclass(frozen=True)
 class DeferredMarshaller(MarshallerABC[T]):
     marshalled_type: T
-    marshaller_context: MarshallerContext
+    marshaller_context: MarshyContext
 
     def get_marshaller(self) -> MarshallerABC[T]:
         marshaller = self.marshaller_context.get_marshaller(self.marshalled_type)
