+++ marshy/marshy_context.py
@@ -11,7 +11,7 @@ from marshy.types import ExternalType
 
 
 @dataclass
-class MarshallerContext:
+class MarshyContext:
     marshallers_by_type: Dict[Type, MarshallerABC] = field(default_factory=dict)
     factories: List[MarshallerFactoryABC] = field(default_factory=list)
 
@@ -51,7 +51,7 @@ class MarshallerContext:
 def marshy_context(injecty_context: Optional[InjectyContext] = None):
     if injecty_context is None:
         injecty_context = get_default_injecty_context()
-    return MarshallerContext(
+    return MarshyContext(
         marshallers_by_type={m.marshalled_type: m for m in injecty_context.get_instances(MarshallerABC)},
         factories=injecty_context.get_instances(MarshallerFactoryABC)
     )
